Move Method to data module and small refactor to avoid circular import
