+++ python/src/pypalmsens/_data/__init__.py
@@ -6,6 +6,7 @@ from .data_value import CurrentReading, PotentialReading
 from .dataset import DataSet
 from .eisdata import EISData
 from .measurement import DeviceInfo, Measurement
+from .method import Material, Method
 from .peak import Peak
 
 __all__ = [
@@ -17,6 +18,8 @@ __all__ = [
     'DeviceInfo',
     'EISData',
     'Measurement',
+    'Material',
+    'Method',
     'Peak',
     'PotentialArray',
     'PotentialReading',
