+++ python/src/pypalmsens/_methods/adapters.py
@@ -1,7 +1,37 @@
 from __future__ import annotations
 
-from pydantic import TypeAdapter
+from typing import Annotated
 
-from . import types
+from pydantic import Field, TypeAdapter
 
-technique_adapter = TypeAdapter(types.TechniqueType)
+from . import mixed_mode, techniques
+
+TechniqueType = Annotated[
+    mixed_mode.MixedMode
+    | techniques.ACVoltammetry
+    | techniques.ChronoAmperometry
+    | techniques.ChronoCoulometry
+    | techniques.ChronoPotentiometry
+    | techniques.CyclicVoltammetry
+    | techniques.DifferentialPulseVoltammetry
+    | techniques.ElectrochemicalImpedanceSpectroscopy
+    | techniques.FastAmperometry
+    | techniques.FastCyclicVoltammetry
+    | techniques.FastGalvanostaticImpedanceSpectroscopy
+    | techniques.FastImpedanceSpectroscopy
+    | techniques.GalvanostaticImpedanceSpectroscopy
+    | techniques.LinearSweepPotentiometry
+    | techniques.LinearSweepVoltammetry
+    | techniques.MethodScript
+    | techniques.MultiStepAmperometry
+    | techniques.MultiStepPotentiometry
+    | techniques.MultiplePulseAmperometry
+    | techniques.NormalPulseVoltammetry
+    | techniques.OpenCircuitPotentiometry
+    | techniques.PulsedAmperometricDetection
+    | techniques.SquareWaveVoltammetry
+    | techniques.StrippingChronoPotentiometry,
+    Field(discriminator='id'),
+]
+
+technique_adapter = TypeAdapter(TechniqueType)
