+++ python/src/pypalmsens/_methods/levels.py
@@ -9,131 +9,6 @@ import PalmSens
 from .._helpers import single_to_double
 from .base_model import BaseModel
 
-AllowedTimingStatus = Literal['Unknown', 'OK', 'OverStep']
-"""Possible values for measurement timing status."""
-
-AllowedReadingStatus = Literal['OK', 'Overload', 'Underload', 'OverloadWarning']
-"""Possible values for current or potential readings."""
-
-AllowedDeviceState = Literal[
-    'Unknown', 'Idle', 'Measurement', 'Download', 'Pretreatment', 'Error', 'MeasOCP'
-]
-"""Possible values for the device state."""
-
-AllowedCurrentRanges = Literal[
-    '100pA',
-    '1nA',
-    '10nA',
-    '100nA',
-    '1uA',
-    '10uA',
-    '100uA',
-    '1mA',
-    '10mA',
-    '100mA',
-    '2uA',
-    '4uA',
-    '8uA',
-    '16uA',
-    '32uA',
-    '63uA',
-    '125uA',
-    '250uA',
-    '500uA',
-    '5mA',
-    '6uA',
-    '13uA',
-    '25uA',
-    '50uA',
-    '200uA',
-    '1A',
-]
-"""Possible current ranges.
-
-See the device documentation or query the instrument manager
-for supported current ranges."""
-
-
-AllowedMethods = Literal[
-    'acv',
-    'ad',
-    'cp',
-    'cpot',
-    'cv',
-    'dpv',
-    'eis',
-    'eis_it',
-    'fam',
-    'fcv',
-    'fgis',
-    'fis',
-    'geis_it',
-    'gis',
-    'gs',
-    'lp',
-    'lsp',
-    'lsv',
-    'ma',
-    'mm',
-    'mp',
-    'mpad',
-    'ms',
-    'npv',
-    'pad',
-    'pot',
-    'ps',
-    'scp',
-    'swv',
-]
-"""All available method IDs.
-
-See the device documentation or query the instrument manager
-for supported methods."""
-
-
-AllowedPotentialRanges = Literal[
-    '1mV',
-    '10mV',
-    '20mV',
-    '50mV',
-    '100mV',
-    '200mV',
-    '500mV',
-    '1V',
-]
-"""Possible potential ranges.
-
-See the device documentation or query the instrument manager
-for supported potential ranges."""
-
-
-def cr_string_to_enum(s: AllowedCurrentRanges) -> PalmSens.CurrentRange:
-    """Convert literal string to CurrentRange."""
-    attr = f'cr{s}'
-    cr = getattr(PalmSens.CurrentRanges, attr)
-
-    return PalmSens.CurrentRange(cr)
-
-
-def cr_enum_to_string(enum: PalmSens.CurrentRange) -> AllowedCurrentRanges:
-    """Convert CurrentRange enum to literal string."""
-    cr = enum.Range
-    return cr.ToString().lstrip('cr')
-
-
-def pr_string_to_enum(s: AllowedPotentialRanges) -> PalmSens.PotentialRange:
-    """Convert literal string to PotentialRange."""
-    attr = f'pr{s}'
-    pr = getattr(PalmSens.PotentialRanges, attr)
-
-    return PalmSens.PotentialRange(pr)
-
-
-def pr_enum_to_string(enum: PalmSens.PotentialRange) -> AllowedPotentialRanges:
-    """Convert PotentialRange enum to literal string."""
-    pr = enum.PR
-    return pr.ToString().lstrip('pr')
-
 
 def convert_bools_to_int(lst: Sequence[bool]) -> int:
     """Convert e.g. [True, False, True, False] to 5."""
@@ -145,53 +20,6 @@ def convert_int_to_bools(val: int) -> tuple[bool, bool, bool, bool]:
     lst = tuple([bool(int(_)) for _ in reversed(f'{val:04b}')])
     assert len(lst) == 4  # specify length to make mypy happy
     return lst
-
-
-def set_extra_value_mask(
-    obj: PalmSens.Method,
-    *,
-    enable_bipot_current: bool = False,
-    record_auxiliary_input: bool = False,
-    record_cell_potential: bool = False,
-    record_dc_current: bool = False,
-    record_we_potential: bool = False,
-    record_forward_and_reverse_currents: bool = False,
-    record_we_current: bool = False,
-):
-    """Set the extra value mask for a given method."""
-    extra_values = 0
-
-    for flag, enum in (
-        (enable_bipot_current, PalmSens.ExtraValueMask.BipotWE),
-        (record_auxiliary_input, PalmSens.ExtraValueMask.AuxInput),
-        (record_cell_potential, PalmSens.ExtraValueMask.CEPotential),
-        (record_dc_current, PalmSens.ExtraValueMask.DCcurrent),
-        (record_we_potential, PalmSens.ExtraValueMask.PotentialExtraRE),
-        (record_forward_and_reverse_currents, PalmSens.ExtraValueMask.IForwardReverse),
-        (record_we_current, PalmSens.ExtraValueMask.CurrentExtraWE),
-    ):
-        if flag:
-            extra_values = extra_values | int(enum)
-
-    obj.ExtraValueMsk = PalmSens.ExtraValueMask(extra_values)
-
-
-def get_extra_value_mask(obj: PalmSens.Method) -> dict[str, bool]:
-    mask = obj.ExtraValueMsk
-
-    ret = {
-        'enable_bipot_current': mask.HasFlag(PalmSens.ExtraValueMask.BipotWE),
-        'record_auxiliary_input': mask.HasFlag(PalmSens.ExtraValueMask.AuxInput),
-        'record_cell_potential': mask.HasFlag(PalmSens.ExtraValueMask.CEPotential),
-        'record_dc_current': mask.HasFlag(PalmSens.ExtraValueMask.DCcurrent),
-        'record_we_potential': mask.HasFlag(PalmSens.ExtraValueMask.PotentialExtraRE),
-        'record_forward_and_reverse_currents': mask.HasFlag(
-            PalmSens.ExtraValueMask.IForwardReverse
-        ),
-        'record_we_current': mask.HasFlag(PalmSens.ExtraValueMask.CurrentExtraWE),
-    }
-
-    return ret
 
 
 class ELevel(BaseModel):
