+++ python/src/pypalmsens/_methods/settings.py
@@ -9,11 +9,13 @@ from typing_extensions import Generator, override
 from .._helpers import single_to_double
 from .base import BaseSettings
 from .base_model import BaseModel
-from .shared import (
-    AllowedCurrentRanges,
-    AllowedPotentialRanges,
+from .levels import (
     convert_bools_to_int,
     convert_int_to_bools,
+)
+from .types import (
+    AllowedCurrentRanges,
+    AllowedPotentialRanges,
     cr_enum_to_string,
     cr_string_to_enum,
     pr_enum_to_string,
