+++ python/src/pypalmsens/settings.py
@@ -2,6 +2,10 @@
 
 from __future__ import annotations
 
+from ._methods.levels import (
+    ELevel,
+    ILevel,
+)
 from ._methods.settings import (
     BiPot,
     BiPotCurrentRange,
@@ -21,15 +25,13 @@ from ._methods.settings import (
     Pretreatment,
     VersusOCP,
 )
-from ._methods.shared import (
+from ._methods.types import (
     AllowedCurrentRanges,
     AllowedDeviceState,
     AllowedMethods,
     AllowedPotentialRanges,
     AllowedReadingStatus,
     AllowedTimingStatus,
-    ELevel,
-    ILevel,
 )
 
 __all__ = [
